+++ CHANGES.md
@@ -1,5 +1,12 @@
 # Changes
 
+## 0.2.0
+
+* fix [issue #7](https://github.com/WZBSocialScienceCenter/geovoronoi/issues/7) (thanks [@AtelierLibre](https://github.com/AtelierLibre))
+* update dependencies / make compatible with GeoPandas 0.7.0
+* introduce more and better tests
+* use *tox* for testing
+
 ## 0.1.2
 
 * fix bug if centroid of boundary area is outside convex hull of generator set (thanks [@mjziebarth](https://github.com/mjziebarth))